Rewrite inlineBearTags based on fromJust
Even though it uses fromJust it probably makes more sense from the
code perspective, as we know that the last element is a tag if
we adjust takeTagText to return a Maybe. Further, this could be
more efficient as we can discard all P.Inline lists which do not
end with a tag string earlier.